Gomathi customer module

+++ OnlineCourierManagementApplication/src/main/java/com/capg/ocma/repository/IComplaintDao.java
@@ -5,9 +5,8 @@ import org.springframework.stereotype.Repository;
 
 import com.capg.ocma.entities.Complaint;
 
-
 @Repository
-public interface IComplaintDao extends JpaRepository<Complaint, Integer> {
+public interface IComplaintDao extends JpaRepository<Complaint, Long> {
